requirements.txt
web3==6.10.0
eth_account
python-dotenv
pandas